Add files via upload
Ten  students  (a,b,c,d,e,f,g,h,i,j)  are  going  to  attend  an  event.  There  are  l
ots  of  gift 
shops, they all are going to the gift shops and randomly picking the gifts. After picking the gifts 
they  are  randomly  arriving  in  the  billing  counter.  The  accountant  gives  the  preference  to  that 
student who has maximum number of gifts. Create a
C or Java program to define order of billed 
students?